Add support for AsteroidOS
AsteroidOS is an open-source operating system for smartwatches [1] built
on Qt 5.6 and OpenEmbedded. This commit adds initial support for
compiling Go Qt programs against the AsteroidOS SDK toolchain. The SDK
includes a script that gets sourced before every build and we use the
OECORE_TARGET_SYSROOT env variable when generating the cgo source files.
Information on building the SDK can be found here [2]. There is
currently no emulator for AsteroidOS so this commit does not add any
code to qtdeploy. Most of the changes are based off the Sailfish code.

Tested using Qt 5.7.0 official prebuilt package and the latest
AsteroidOS SDK.

[1] https://asteroidos.org/
[2] https://asteroidos.org/wiki/creating-an-asteroid-app/