Fix for bundling when using MSYS2 outside of MSYS2 Shell
When using a setup where you are outside of MSYS2 shell, but you still want to use the MSYS2 build of Qt binaries for 64bit.  You can see the configuration here https://github.com/rashwell/neochess/wiki/NeoChess-Building-on-Windows-with-MSYS2,-VSCode,-and-64-bit-Qt the build and bundle deploy works fine, but without the /Y switch to the xcopy in the patch here, the bundle process displays errors.